requirements.txt
python3 manage.py makemigrations
python3 manage.py migrate
python3 manage.py createsuperuser
python3 manage.py runserver
API endpoints at http://127.0.0.1:8000/api/


appdirs==1.4.4
apturl==0.5.2
asgihandler==0.5.9
asgiref==3.2.10
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
base58==1.0.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cbor==1.0.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==40.0.2
cssselect2==0.4.1
cupshelpers==1.0
Cython==0.29.28
daphne==4.0.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.0.8
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
docopt==0.6.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
duplicity==0.8.21
ecdsa==0.18.0b1
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.13.1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gyp==0.1
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httplib2==0.20.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
mnemonic==0.19
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
netifaces==0.11.0
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.13
platformdirs==4.1.0
ply==3.11
priority==1.3.0
protobuf==3.12.4
ptyprocess==0.7.0
py-ubjson==0.16.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydyf==0.1.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyperclip==1.8.2
pyphen==0.12.0
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-snappy==0.5.3
pythran==0.10.0
PyTrie==0.4.0
pytz==2020.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
request2==0.2
requests==2.28.1
scipy==1.8.0
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
sortedcontainers==2.1.0
soupsieve==2.3.1
sqlparse==0.3.1
sympy==1.9
systemd-python==234
tinycss2==1.1.1
ttystatus==0.38
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2022.5
u-msgpack-python==2.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unicodecsv==0.14.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
weasyprint==54.1
webencodings==0.5.1
wsaccel==0.6.3
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
zope.interface==6.0
